+++ docs/source/reference/internals.rst
@@ -4,11 +4,14 @@ Internal API
 MaskContainer
 ~~~~~~~~~~~~~
 
-:class:`libertem.job.masks.MaskContainer` helps to implement highly efficient
+:class:`libertem.common.container.MaskContainer` helps to implement highly efficient
 mask application operations, such as virtual detector, center of mass or feature
 vector calculations.
 
-.. automodule:: libertem.job.masks
+.. versionchanged:: 0.4.0.dev0
+    Moved from :mod:`libertem.job.masks` to :mod:`libertem.common.container` to use it in UDFs and prepare depreciation of the Job interface.
+
+.. automodule:: libertem.common.container
    :members: MaskContainer
    :special-members: __init__
